Fix OpenGL texture leak on eviction, 64-bit undo accounting, container memsize, and deferred visual validation
- Explicitly release gl_tid via unloadGLTexture(0) in GameSprite::TemplateImage and NormalImage destructors before base destruction, ensuring textures are deleted on LRU eviction without double-decrementing loaded_textures
- Expand Action, Change, BatchAction, Item, Container, and Tile memory accounting paths to size_t with saturating addition, preventing 32-bit overflow on large graphs
- Include sizeof(contents) in Container::memsize() to account for inline vector storage in addition to heap capacity and nested items
- Validate deferred appearance metadata (requiring drawable frame group and resolved first sprite sheet) in GraphicManager and cache failures, preventing unusable visuals from reporting as available
- Add regression and boundary test coverage in spell_visual_browser_memory_tests and editor_data_layout_tests

+++ source/complexitem.h
@@ -51,7 +51,7 @@ public:
 		return contents;
 	}
 	double getWeight() const override;
-	uint32_t memsize() const override;
+	size_t memsize() const override;
 
 	bool unserializeItemNode_OTBM(const IOMap& maphandle, BinaryNode* node) override;
 	bool serializeItemNode_OTBM(const IOMap& maphandle, NodeFileWriteHandle& f) const override;
